Install dependencies, set up react-router-dom, eslint config, split into separate folders
