+++ .eslintrc.js
@@ -1,0 +1,23 @@
+module.exports = {
+  env: {
+    browser: true,
+    commonjs: true,
+    es6: true,
+    jest: true,
+    node: true
+  },
+  extends: ["eslint:recommended", "plugin:react/recommended"],
+  globals: {
+    Atomics: "readonly",
+    SharedArrayBuffer: "readonly"
+  },
+  parserOptions: {
+    sourceType: "module",
+    ecmaFeatures: {
+      jsx: true
+    },
+    ecmaVersion: 2018
+  },
+  plugins: ["react"],
+  rules: {}
+};
