--- src/App.js
@@ -1,12 +1,0 @@
-import React from "react";
-import "./App.css";
-
-const App = () => {
-  return (
-    <>
-      <h1>Gateway 2 Heritage</h1>
-    </>
-  );
-};
-
-export default App;
