+++ src/assets/assets.txt
@@ -1,0 +1,1 @@
+placeholder txt
