+++ src/components/App/App.js
@@ -1,0 +1,19 @@
+import React from "react";
+import "./App.css";
+// import { BrowserRouter, Route, Switch, Link } from "react-router-dom";
+
+const App = () => {
+  return (
+    // <BrowserRouter>
+    //   <Switch>
+    <>
+      <h1>Gateway 2 Heritage</h1>
+      {/* <Route path="/" component={*component*} exact*/}
+      {/* <Route path="/" render=(() => <Component/>) */}
+    </>
+    //  </Switch>
+    // </BrowserRouter>
+  );
+};
+
+export default App;
